# [symbols] Change format of symbols.sdf