… data/resources-6plus_dark/symbols.sdf
… data/resources-6plus_light/symbols.sdf
… data/resources-hdpi_dark/symbols.sdf
… data/resources-hdpi_light/symbols.sdf
… data/resources-mdpi_dark/symbols.sdf
… data/resources-mdpi_light/symbols.sdf
… data/resources-xhdpi_dark/symbols.sdf
… data/resources-xhdpi_light/symbols.sdf
… data/resources-xxhdpi_dark/symbols.sdf
… data/resources-xxhdpi_light/symbols.sdf
… data/resources-xxxhdpi_dark/symbols.sdf
… data/resources-xxxhdpi_light/symbols.sdf