Changed format from xml to csv
This will save us ~0.5 mb and speed up reading a bit

| | Old | New | Diff |
|--------|--------|--------|--------|
| M1 Pro (debug) | 3917417 | 2584750 | 1332667 (1.3 ms)
| Old Android (debug) | 22607615 | 5583077 | 17024538 (17 ms)
| Old Android (release) | 1964615 | 976385 | 988230 (1 ms)


I don't really like the new implementation. Could it be simplified? Do we have csv reader implemented somewhere?