@biodranik 2025-07-04
json + https://stephenberry.github.io/glaze/ ?