@AndrewShkrob 2025-09-25
glaze should be better here than csv :)